Added multicast-relay service and default settings

+++ misc/upgrade-helper.sh
@@ -138,10 +138,8 @@ chmod +x /home/{pi,vpn}/.bash* /etc/skel/{.bash*,.profile}
 #####################################################################################
 for DEST in $(cat $TFL); do 
 	[[ "${QUIET}" == "false" ]] && echo -e "Removing ${BLUE}${DEST}${NC}... "
-	if [[ "$DEST" =~ /etc/systemd/system/ ]]; then
-		systemctl disable --now $(basename $DEST)
-	fi
-	rm ${DEST}
+	[[ "$DEST" =~ /(lib|etc)/systemd/system/ ]] && systemctl disable --now $(basename $DEST)
+	test -f ${DEST} && rm ${DEST}
 done
 
 #####################################################################################
@@ -157,6 +155,7 @@ if [[ ! -z "${RW[5]}" ]]; then
 	systemctl is-enabled cloudflared@1 >& /dev/null || systemctl enable ${NOW} cloudflared@1
 	systemctl is-enabled cloudflared@2 >& /dev/null || systemctl enable ${NOW} cloudflared@2
 	systemctl is-enabled cloudflared@3 >& /dev/null || systemctl enable ${NOW} cloudflared@3
+	systemctl is-enabled multicast-relay >& /dev/null || systemctl enable ${NOW} multicast-relay
 	systemctl is-enabled wifi >& /dev/null || systemctl enable ${NOW} wifi
 	chroot /ro /opt/bpi-r2-router-builder/upgrade.sh -f
 	[[ "${RW[5]}" == *ro,* ]] && mount -o remount,ro /ro
